# health_check filter: optimize predicate by eliminating floating point…