This PR fixes a TODO in the health check filter by translating the floating point arithmetic to integer arithmetic instead.

The performance gain is minimal. I wrote a bench test where I isolate the predicate in the if statement to compare the two implementations:

The results (with `--benchmark_repetitions=100`):
BM_IsHealthyOrig_mean         4.53 ns
BM_IsHealthyOpt_mean          4.17 ns

I also adapted one of the test cases from the health check filter test suite and made it into a bench test so I can exercise the filter code itself.

The results (with `--benchmark_repetitions=100`):
BM_DecodeHeaders_mean        60887 ns  (existing code)
BM_DecodeHeaders_mean        60682 ns  (optimized)

The bench test can be found on this branch:
https://github.com/numerodix/envoy/blob/8657a2bdf5511cff1a6db58bc35631e5b5570f95/test/extensions/filters/http/health_check/health_check_speed_test.cc

I also compared the two predicate implementations in https://godbolt.org/ and the optimized version (x86-64/clang-10 with -O3) produces fewer instructions. The difference can be seen here:
existing: https://github.com/numerodix/envoy/blob/8657a2bdf5511cff1a6db58bc35631e5b5570f95/orig#L20
optimized: https://github.com/numerodix/envoy/blob/8657a2bdf5511cff1a6db58bc35631e5b5570f95/opt#L9

Risk level: Medium (modification of existing filter)
Testing: Re-ran the health check filter tests. This is a refactoring and the existing test suite has several test cases exercising this logic, so I did not add any new ones.

Signed-off-by: Martin Matusiak <numerodix@gmail.com>