@snowp 2020-06-18
Actually one question: given that this is a minor change in behavior, does this warrant a runtime flag?