@numerodix 2020-06-19
I would think it's a gray area.

from CONTRIBUTING.md:
> Some changes in Envoy are deemed worthy of runtime guarding. Instead of just replacing
old code with new code, both code paths are supported for between one Envoy release (if it is
guarded due to performance concerns) and a full deprecation cycle (if it is a high risk behavioral
change).

It's a medium risk behavioral change, so I'm not sure if it's warranted.

If we do want this then I would move the predicate in the if statement to a separate function, and do the runtime guard there. It would incur doing the check every time the if statement is executed, so that would potentially negate the benefit of this optimization during the transition period.